docs: rename SourceLinks.vue to EssayMeta.vue
The component renders both the draft/speculative status banner and the
"Source files" box, so the old name described only half its job.
